+++ src/Mod/Mesh/Gui/ViewProviderMeshFaceSet.cpp
@@ -124,7 +124,7 @@ ViewProviderMeshFaceSet::ViewProviderMeshFaceSet() : pcOpenEdge(0), m_bEdit(fals
     unsigned long current = color.getPackedValue();
     unsigned long setting = hGrp->GetUnsigned("MeshColor", current);
     if (current != setting) {
-        color.setPackedValue(setting);
+        color.setPackedValue((uint32_t)setting);
         ShapeColor.setValue(color);
     }
